// learning-react/src/models/Employee.ts
export interface Employee {
    id: string;
    uname: string;
    age: number;
    gender: string;
    dept: string;
    doj: string;
    salary: number;
    cont: string;
    email: string;
    address: string;
    skills: string[];
    employmentType: string;
    experience: number;
    active: boolean;
}